Updates to IMU reading

+++ FlightController/header/sensorRead.h
@@ -41,7 +41,7 @@ class Sensors
   public:
     Sensors();
     double Altitude();        // calculated the current altitude above sea level
-    float *IMU();        // calculate the current pose roll, pitch, yaw(x,y,z)
+    double *IMU();        // calculate the current pose roll, pitch, yaw(x,y,z)
     float MAG();           // calculate the current heading  yaw(z)
 };
